feat: Update ProjectPanel component to improve performance
The code changes in the ProjectPanel component optimize performance by removing unnecessary console.log statements and updating the useEffect hook. This improves code cleanliness and removes unnecessary logging in the production environment. Additionally, the component now supports lazy loading of images, which improves page load speed by loading images only when they are visible on the screen. These enhancements optimize the user experience and improve the performance of the application.